filter_exclude: add new audit filter exclude tests
https://github.com/linux-audit/audit-kernel/issues/5

In addition to the originally supported MSGTYPE, test for support for
PID, UID, GID, LOGINUID, (LOGINUID_SET,) subj_user, subj_role,
subj_type, subj_sen, subj_clr and verify that a sampling of others are
not supported.

Since the exclude filter is checked before the exit filter, use a
syscall (open, unlink) filter to verify the messages are blocked.

Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
[PM: adjusted patch subject line and minor merge fixes]
Signed-off-by: Paul Moore <paul@paul-moore.com>