all: introduce an environment variable for additional debug output
Document the "ATS_DEBUG" environment variable which can be used in
individual tests to trigger additional debug output.  At the moment
we are only specifying values of not-set, 0, and 1; with the value 1
being the only value which would trigger the additional output.  We
reserve the right to add additional recognized values in the future
for varying degrees of verbosity (e.g. ATS_DEBUG=2).

Sample Perl code to check if ATS_DEBUG is set to 1:

  if (defined $ENV{ATS_DEBUG} && $ENV{ATS_DEBUG} == 1) {
          # debug output
  }

Signed-off-by: Paul Moore <paul@paul-moore.com>