audit lost reset: reset backlog_wait_time and backlog_limit on cleanup
Since restarting the daemon does not reset these two values unless they
are explicitly set in the config file, reset them manually based on the
auditctl status values at the start of the test.  While it would be
better to use the actual kernel defaults, those are not available to
this test.

See:
	https://github.com/linux-audit/audit-kernel/issues/3
	https://github.com/linux-audit/audit-testsuite/issues/56
	https://github.com/linux-audit/audit-testsuite/pull/41
Signed-off-by: Richard Guy Briggs <rgb@redhat.com>