tests: prevent ausearch race conditions
On busy machines it can happen that the record we are searching for
doesn't make it to the log file before we run ausearch.

This patch fixes this problem by adding a short snippet to the sync
points, which generates a marker user record using 'auditclt -m' and
waits until the marker appears in the audit log. This is much more
reliable than the current mix of either completely ignoring the issue or
just doing an unreliable sleep.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>