audit-testsuite: try harder on the lost_reset test
If our attempts at causing lost records fail, bring out the "big
guns" and try a bit harder.  Because of this ability to hit the
system a bit harder, we also adjust the number of iterations in
an attempt to speed things up if the system is not dropping
records using the easy approach.

This change also requires us to add a
small sleep before restarting auditd at the end of the test as
systemd was complaining about the restarts and failing to (re)start
auditd.

Suggested-by: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>