tests/filter_exclude: fix MLS range parsing
If the test is run under a context with identical lower and upper MLS
levels, then the context uses a simplified format with only one level
and no "-". This confuses the filter_exclude test, which assumes the
range is always of the form "<lower>-<upper>".

Fix the regex to treat the upper level as optional and set it to the
lower one in that case. Note that we also need to use the original $subj
string in the ausearch query, since ausearch wouldn't match the
"...:<level>-<level>" context to "...:<level>".

Resolves: https://github.com/linux-audit/audit-testsuite/issues/58
Fixes: 447acf74c4f8 ("filter_exclude: add new audit filter exclude tests")
Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>