io_uring: add initial io_uring tests
This is rather simple test using a basic set of io_uring operations
to open a file, write a string to it, read the string back out, and
then close the file.  The test verifies that the io_uring related
records are captured and marked with a given key value and that the
uring operation values are present.

Signed-off-by: Paul Moore <paul@paul-moore.com>