filter_exclude: remove auid check from unlink SYSCALL mesage check
The test filter_exclude fails on cases 20 and 21 if you log on the test
machine using a non-root user and then switch to root.
Removing auid check to avoid this failure, since the auid check is not
required in the context of these test cases.

See: https://github.com/linux-audit/audit-testsuite/issues/113

Signed-off-by: Ricardo Robaina <rrobaina@redhat.com>